Moved add services logic to separate component
